+++ pkg/distributor/validate.go
@@ -43,6 +43,7 @@ var (
 	reasonDuplicateLabelNames          = globalerror.SeriesWithDuplicateLabelNames.LabelValue()
 	reasonTooFarInFuture               = globalerror.SampleTooFarInFuture.LabelValue()
 	reasonTooFarInPast                 = globalerror.SampleTooFarInPast.LabelValue()
+	reasonDuplicateTimestamp           = globalerror.SampleDuplicateTimestamp.LabelValue()
 
 	// Discarded exemplars reasons.
 	reasonExemplarLabelsMissing               = globalerror.ExemplarLabelsMissing.LabelValue()
@@ -98,6 +99,7 @@ var (
 		"received a sample whose timestamp is too far in the past, timestamp: %d series: '%.200s'",
 		validation.PastGracePeriodFlag,
 	)
+	duplicateTimestampMsgFormat  = globalerror.SampleDuplicateTimestamp.Message("samples with duplicated timestamps have been discarded, discarded samples: %d series: '%.200s'")
 	exemplarEmptyLabelsMsgFormat = globalerror.ExemplarLabelsMissing.Message(
 		"received an exemplar with no valid labels, timestamp: %d series: %s labels: %s",
 	)
@@ -143,6 +145,7 @@ type sampleValidationMetrics struct {
 	duplicateLabelNames          *prometheus.CounterVec
 	tooFarInFuture               *prometheus.CounterVec
 	tooFarInPast                 *prometheus.CounterVec
+	duplicateTimestamp           *prometheus.CounterVec
 }
 
 func (m *sampleValidationMetrics) deleteUserMetrics(userID string) {
@@ -160,6 +163,7 @@ func (m *sampleValidationMetrics) deleteUserMetrics(userID string) {
 	m.duplicateLabelNames.DeletePartialMatch(filter)
 	m.tooFarInFuture.DeletePartialMatch(filter)
 	m.tooFarInPast.DeletePartialMatch(filter)
+	m.duplicateTimestamp.DeletePartialMatch(filter)
 }
 
 func (m *sampleValidationMetrics) deleteUserMetricsForGroup(userID, group string) {
@@ -176,6 +180,7 @@ func (m *sampleValidationMetrics) deleteUserMetricsForGroup(userID, group string
 	m.duplicateLabelNames.DeleteLabelValues(userID, group)
 	m.tooFarInFuture.DeleteLabelValues(userID, group)
 	m.tooFarInPast.DeleteLabelValues(userID, group)
+	m.duplicateTimestamp.DeleteLabelValues(userID, group)
 }
 
 func newSampleValidationMetrics(r prometheus.Registerer) *sampleValidationMetrics {
@@ -193,6 +198,7 @@ func newSampleValidationMetrics(r prometheus.Registerer) *sampleValidationMetric
 		duplicateLabelNames:          validation.DiscardedSamplesCounter(r, reasonDuplicateLabelNames),
 		tooFarInFuture:               validation.DiscardedSamplesCounter(r, reasonTooFarInFuture),
 		tooFarInPast:                 validation.DiscardedSamplesCounter(r, reasonTooFarInPast),
+		duplicateTimestamp:           validation.DiscardedSamplesCounter(r, reasonDuplicateTimestamp),
 	}
 }